requirements.txt
streamlit
openai
whisper
gtts
python-dotenv
